Development of the form view to create a new article
The call to the route corresponding to the articles was established

resources: articles

The controller was created to generate the articles

$ rails g controller articles

The new method (for now empty) was defined, which allows to visualize the screen where the view will be.

The view is created to create a new article: new.html.erb, in which a Rails form builder is used.

The visual appearance of that form was established through an SCSS file.

In the articles controller, a method called create (empty for now) was defined, so as not to generate an error when submitting from the form.

Task:

Generate the database

Perform in the controller, the option that can save the new articles in the database.