The method was created to save the elements in the database
Inside the Articles controller, in the create method, the code was established so that the information is stored in the database established in the migrations.

A private type method was created to allow access to the database parameters.

The show method was created, where the elements of the database are called after saving them.

The view was created to show the new items being saved.

show.html.erb