Show posts made
The corresponding codes are added to the Articles controller and to the index.html.erb view so that the list of publications made in a table appears.

The corresponding links have been added to return to the listing page from a new article and show article.

Validations were added, to prevent the user from entering empty data in the form, in the view of new.html.erb.

Rubocop was used to correct some visual code errors.