Updating of publications
The edit and update methods were created in the Articles controller, in order to update published articles.

A view called: edit.html.erb was created, which allows to generate a form calling the information of the Articles controller and to be able to edit the information.

In the index.html.erb view, two buttons were added, one to display the article, and one to edit it.

A new view called _form.html.erb was created to eliminate duplications in the new and edit views, through a partial.