Create comments
A controller for comments was generated, through the following code:

$ rails generate controller Comments

The create function was set in the controller to allow the creation of new comments in the database.

Two partials were created: one to show the comments made and the other to show the form with which to create the comment.