Basic authentication
Access protection is created for several of the actions defined in the ArticlesController controller. If the user is not authenticated, they will not be able to access those actions. For this we use the http_basic_authenticate_with method of Rails.

Another additional restriction is preventing users from deleting comments.